+++ src/View/Components/CopyUrl.php
@@ -11,7 +11,7 @@ class CopyUrl extends Component
 
     public function __construct(
         ?string $url = null,
-        public string $icon = 'fa-solid fa-fw fa-clipboard',
+        public string $icon = 'fa-regular fa-fw fa-copy',
         public string $copiedIcon = 'fa-solid fa-fw fa-check',
     ) {
         $this->url = $url ?? url()->full();